+++ apps/playground/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @ai-gui/playground
 
+## 0.0.5
+
+### Patch Changes
+
+- @ai-gui/core@0.4.3
+- @ai-gui/react@0.4.3
+- @ai-gui/vue@0.4.3
+- @ai-gui/vanilla@0.4.3
+- @ai-gui/plugin-mermaid@0.4.3
+- @ai-gui/plugin-citation@0.4.3
+- @ai-gui/plugin-artifact@0.4.3
+- @ai-gui/plugin-ui@0.4.3
+- @ai-gui/plugin-molecule@0.4.3
+- @ai-gui/plugin-map@0.4.3
+- @ai-gui/devtools@0.4.3
+
 ## 0.0.4
 
 ### Patch Changes